% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_named_list}
\alias{auto_named_list}
\title{Creates a named list where the names are taken from the input variables}
\usage{
auto_named_list(...)
}
\arguments{
\item{...}{list elements}
}
\value{
named list of elements
}
\description{
NOTE: only supports arguments, not keyword arguments
}
\examples{
a <- 1
b <- 2
auto_named_list(a, b)$a == 1
auto_named_list(a, b)$b == 2
}
