% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_slise}
\alias{create_slise}
\title{Create a result object for SLISE that is similar to other regression method results}
\usage{
create_slise(alpha, X, Y, epsilon, lambda = 0, data = NULL, ...)
}
\arguments{
\item{alpha}{linear model}

\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{lambda}{L1 weight}

\item{data}{data_preprocess(X, Y)}

\item{...}{other variables to add to the SLISE object}
}
\value{
list(coefficients=unscale(alpha), X, Y, scaled=data, lambda, alpha, subset=[r_i<epsilon], value=loss, epsilon, loss, ...)
}
\description{
Create a result object for SLISE that is similar to other regression method results
}
