% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{data_container}
\alias{data_container}
\title{Wrapper for creating a C++ DataContainer that parses parameter names}
\usage{
data_container(X, Y, epsilon, beta, lambda1 = 0, lambda2 = 0, weight = NULL)
}
\arguments{
\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{beta}{sigmoid steepness}

\item{lambda1}{L1 regularisation (default: 0)}

\item{lambda2}{L2 regularisation (default: 0)}

\item{weight}{weight vector (default: NULL)}
}
\value{
DataContainer
}
\description{
Wrapper for creating a C++ DataContainer that parses parameter names
}
