% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_identity}
\alias{data_identity}
\title{SLISE Data Preprocessing without changing anything}
\usage{
data_identity(X, Y)
}
\arguments{
\item{X}{the data matrix}

\item{Y}{the response vector}
}
\value{
list(X, Y, scaling_functions...)
}
\description{
SLISE Data Preprocessing without changing anything
}
