% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_local}
\alias{data_local}
\title{Shift the data to be local around a point}
\usage{
data_local(X, Y, x, y = NULL)
}
\arguments{
\item{X}{data matrix}

\item{Y}{response vector}

\item{x}{local data point}

\item{y}{local response}
}
\value{
list(X = X - x, Y = Y - y, scaling_functions...)
}
\description{
Shift the data to be local around a point
}
