% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_preprocess}
\alias{data_preprocess}
\title{SLISE Data Preprocessing
Scales data and optionally adds intercept}
\usage{
data_preprocess(X, Y, scale = FALSE, intercept = FALSE,
  logit_tr = FALSE, scale_y = TRUE)
}
\arguments{
\item{X}{the data matrix}

\item{Y}{the response vector}

\item{scale}{should the data be (columnwise) normalised (FALSE)}

\item{intercept}{should an intercept column be added (FALSE)}

\item{logit_tr}{should the response be logit-transformed (FALSE)}

\item{scale_y}{should the response be scaled (FALSE)}
}
\value{
list(X, Y, scaling_functions...)
}
\description{
SLISE Data Preprocessing
Scales data and optionally adds intercept
}
