% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dlog_sigmoid}
\alias{dlog_sigmoid}
\title{derivative of log-sigmoid function}
\usage{
dlog_sigmoid(x)
}
\arguments{
\item{x}{vector of real values}
}
\value{
Derivative of log(sigmoid(x)).
}
\description{
derivative of log-sigmoid function
}
