% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dsigmoid}
\alias{dsigmoid}
\title{derivative of sigmoid function}
\usage{
dsigmoid(x)
}
\arguments{
\item{x}{vector of real values}
}
\value{
Derivative of sigmoid(x).
}
\description{
derivative of sigmoid function
}
