% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{explain}
\alias{explain}
\title{Explain an Object}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{the object to explain}

\item{...}{additional parameters}
}
\description{
Explain an Object
}
\examples{
data <- matrix(rnorm(200), 100, 2)
response <- rnorm(100)
slise <- slise.explain(data, response, 10, epsilon=0.1)
explain(slise)
}
