% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{explain.slise}
\alias{explain.slise}
\title{Show a SLISE explanation}
\usage{
\method{explain}{slise}(slise, type = "bar",
  class_labels = c("Class 0", "Class 1"), title = "SLISE",
  real_value = NULL, probability = TRUE, ...)
}
\arguments{
\item{slise}{the slise object to show}

\item{type}{Type of explanation (bar, distribution, image, image2, text, image_scatter)}

\item{class_labels}{(Optional) names of the two classes}

\item{title}{(Optional) A title to add to the explanation}

\item{real_value}{(Optional) the real response for the explained instance}

\item{probability}{(Optional) is The prediction a probability (TRUE)}

\item{...}{Additional parameters to the visualisation}
}
\description{
Show a SLISE explanation
}
\examples{
data <- matrix(rnorm(200), 100, 2)
response <- rnorm(100)
slise <- slise.explain(data, response, 10, epsilon=0.1)
explain(slise)
}
