% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{fast_ols}
\alias{fast_ols}
\title{OLS solver that falls back to an optimisation if ncol(X) is huge
Also supports LASSO via optimisation}
\usage{
fast_ols(X, Y, weight = NULL, lambda = 0, max_iterations = 300)
}
\arguments{
\item{X}{data matrix}

\item{Y}{response vector}

\item{weight}{weight vector (default: NULL)}

\item{lambda}{LASSO regularisation (default: 0)}

\item{max_iterations}{if ncol(X) is huge, then ols is replaced with optimisation (default:300)}
}
\value{
coefficient vector
}
\description{
OLS solver that falls back to an optimisation if ncol(X) is huge
Also supports LASSO via optimisation
}
