% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limited_logit}
\alias{limited_logit}
\title{Computes the logits from probabilities}
\usage{
limited_logit(p, stab = 0.001)
}
\arguments{
\item{p}{probability (vector)}

\item{stab}{limit p to [stab, 1-stab] for numerical stability}
}
\value{
log(p / (1 - p))
}
\description{
Computes the logits from probabilities
}
