% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{log_approximation_ratio}
\alias{log_approximation_ratio}
\title{Calculate the Logarithm of the approximation ratio
(logarithms are used for numerically stable calculations)
See Theorem 3 from the paper for more details}
\usage{
log_approximation_ratio(residuals2, epsilon2, beta1, beta2, weight = NULL)
}
\arguments{
\item{residuals2}{squared residuals}

\item{epsilon2}{squared error tolerance}

\item{beta1}{current sigmoid steepness}

\item{beta2}{next sigmoid steepness}

\item{weight}{weight vector (default: NULL)}
}
\value{
log(approximation_ratio)
}
\description{
Calculate the Logarithm of the approximation ratio
(logarithms are used for numerically stable calculations)
See Theorem 3 from the paper for more details
}
