% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum}
\alias{log_sum}
\title{Computes log(sum(exp(x))) in a numerically robust way.}
\usage{
log_sum(x)
}
\arguments{
\item{x}{vector of length n}
}
\value{
log(sum(exp(x))).
}
\description{
Computes log(sum(exp(x))) in a numerically robust way.
}
