% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_special}
\alias{log_sum_special}
\title{Computes log(sum(exp(x) * y)),
or log(sum(exp(x))) if all(y == 0),
in a numerically robust way.}
\usage{
log_sum_special(x, y)
}
\arguments{
\item{x}{vector of length n}

\item{y}{multiplier}
}
\value{
log(sum(exp(x))).
}
\description{
Computes log(sum(exp(x) * y)),
or log(sum(exp(x))) if all(y == 0),
in a numerically robust way.
}
