% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_sharp}
\alias{loss_sharp}
\title{Sharp Loss Function}
\usage{
loss_sharp(alpha, X, Y, epsilon = 0.1, lambda = 0)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{X}{The data matrix}

\item{Y}{The response vector}

\item{epsilon}{(Optional) The acceptable error}

\item{lambda}{(Optional) The sparsity reguraliser}
}
\value{
The loss value
}
\description{
Sharp Loss Function
}
