% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_sharp_res}
\alias{loss_sharp_res}
\title{Sharp Loss Function
Exact loss function without gradients for when the residuals are already calculated}
\usage{
loss_sharp_res(
  alpha,
  residuals2,
  epsilon2,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL
)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{residuals2}{The squared error vector: (X \%*\% alpha - Y)^2}

\item{epsilon2}{The squared acceptable error}

\item{lambda1}{The L1 regulariser (default: 0)}

\item{lambda2}{The L2 regulariser (default: 0)}

\item{weight}{weight vector (default: NULL)}
}
\value{
The loss value
}
\description{
Sharp Loss Function
Exact loss function without gradients for when the residuals are already calculated
}
