% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_smooth}
\alias{loss_smooth}
\title{Smooth Loss}
\usage{
loss_smooth(alpha, X, Y, epsilon = 0.1, lambda = 0, beta = 3)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{X}{The data matrix}

\item{Y}{The response vector}

\item{epsilon}{(Optional) The acceptable error}

\item{lambda}{(Optional) The sparsity reguraliser}

\item{beta}{(Optional) The steepness of the sigmoid (default: 3)}
}
\value{
The loss value
}
\description{
Smooth Loss
}
