% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_smooth_res}
\alias{loss_smooth_res}
\title{Smooth Loss
A loss function for when you want gradients and the residuals are already calculated}
\usage{
loss_smooth_res(
  alpha,
  residuals2,
  epsilon2,
  beta,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL
)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{residuals2}{Vector of squared residuals}

\item{epsilon2}{The squared acceptable error}

\item{beta}{The steepness of the sigmoid}

\item{lambda1}{The L1 regulariser (default: 0)}

\item{lambda2}{The L2 regulariser (default: 0)}

\item{weight}{weight vector (default: NULL)}
}
\value{
The loss value
}
\description{
Smooth Loss
A loss function for when you want gradients and the residuals are already calculated
}
