% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{matching_epsilon}
\alias{matching_epsilon}
\title{Find the matching *epsilon}
\usage{
matching_epsilon(residuals, epsilon, beta)
}
\arguments{
\item{residuals}{squared residuals}

\item{epsilon}{error tolerance}

\item{beta}{sigmoid steepness}
}
\value{
*epsilon
}
\description{
Find the matching *epsilon
}
