% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{owlqn_c}
\alias{owlqn_c}
\title{OWL-QN for optimising loss_smooth
Cpp implementation}
\usage{
owlqn_c(alpha, dc, lambda1 = 0, max_iterations = 300, ..., invisible = TRUE)
}
\arguments{
\item{alpha}{linear model to optimise}

\item{dc}{DataContainer containing the data and parameters}

\item{lambda1}{L1 coefficient (default: 0)}

\item{max_iterations}{number of OWL-QN iterations (default: 300)}

\item{...}{other parameters to OWL-QN}

\item{invisible}{no terminal output (default: TRUE)}
}
\value{
lbfgs object
}
\description{
OWL-QN for optimising loss_smooth
Cpp implementation
}
