% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{owlqn_r}
\alias{owlqn_r}
\title{OWL-QN for optimising loss_smooth (R implementation)}
\usage{
owlqn_r(alpha, X, Y, epsilon = 0.1, lambda = 0, beta = 3,
  max_iterations = 250, ...)
}
\arguments{
\item{alpha}{linear model to optimise}

\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{lambda}{L1 coefficient}

\item{beta}{sigmoid steepness}

\item{max_iterations}{number of OWL-QN iterations}

\item{...}{other parameters to OWL-QN}
}
\value{
lbfgs object
}
\description{
OWL-QN for optimising loss_smooth (R implementation)
}
