% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise_2d}
\alias{plot.slise_2d}
\title{Plot the robust regression or explanation from slise in 2D}
\usage{
\method{plot}{slise_2d}(slise, title = NULL, labels = NULL, partial = FALSE, size = 2, ...)
}
\arguments{
\item{slise}{The slise object}

\item{title}{The title of the plot (may include a `\%s`, which will be replaced by the prediction)}

\item{labels}{The axis labels (default: c("X", "Y") or c("x", "f(x)"))}

\item{partial}{Should the raw ggplot2 objects be returned instead of directly plotting (default: FALSE)}

\item{size}{The size of the plot elements (default: 2)}

\item{...}{Ignored parameters}
}
\value{
ggplot object or plot
}
\description{
Plot the robust regression or explanation from slise in 2D
}
