% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise_bar}
\alias{plot.slise_bar}
\title{Plot the robust regression or explanation from slise as bar plots}
\usage{
\method{plot}{slise_bar}(
  slise,
  title = NULL,
  labels = c("Low", "High"),
  partial = FALSE,
  size = 8,
  ...
)
}
\arguments{
\item{slise}{The slise object}

\item{title}{The title of the plot (may include a `\%s`, which will be replaced by the prediction)}

\item{labels}{The class labels (vector with two strings: c(y_low, y_high), default: c("Low", "High"))}

\item{partial}{Should the raw ggplot2 objects be returned instead of directly plotting (default: FALSE)}

\item{size}{The size of the segments (default: 8)}

\item{...}{Ignored parameters}
}
\value{
List of ggplot objects or plot
}
\description{
Plot the robust regression or explanation from slise as bar plots
}
