% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise_distribution}
\alias{plot.slise_distribution}
\title{Plot the robust regression or explanation from slise with distributions}
\usage{
\method{plot}{slise_distribution}(
  slise,
  title = NULL,
  labels = c("Low", "High"),
  partial = FALSE,
  signif = 3,
  ...
)
}
\arguments{
\item{slise}{The slise object}

\item{title}{The title of the plot (may include a `\%s`, which will be replaced by the prediction)}

\item{labels}{The class labels (vector with two strings: c(y_low, y_high), default: c("Low", "High"))}

\item{partial}{Should the raw ggplot2 objects be returned instead of directly plotting (default: FALSE)}

\item{signif}{The number of significant digits to display (default: 3)}

\item{...}{Ignored parameters}
}
\value{
List of ggplot objects or plot
}
\description{
Plot the robust regression or explanation from slise with distributions
}
