% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise_prediction}
\alias{plot.slise_prediction}
\title{Plot the robust regression or explanation from slise based on predictions}
\usage{
\method{plot}{slise_prediction}(
  slise,
  title = NULL,
  labels = c("Response", "Count"),
  partial = FALSE,
  approximation = TRUE,
  signif = 3,
  ...
)
}
\arguments{
\item{slise}{The slise object}

\item{title}{The title of the plot (may include a `\%s`, which will be replaced by the prediction)}

\item{labels}{The axis labels (default: c("Response", "Count"))}

\item{partial}{Should the raw ggplot2 objects be returned instead of directly plotting (default: FALSE)}

\item{approximation}{Should the approximation density be added (default: TRUE)}

\item{signif}{The number of significant digits to display (default: 3)}

\item{...}{Ignored parameters}
}
\value{
ggplot object or plot
}
\description{
Plot the robust regression or explanation from slise based on predictions
}
