% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise_wordcloud}
\alias{plot.slise_wordcloud}
\title{Plot the robust regression or explanation from slise as a wordcloud}
\usage{
\method{plot}{slise_wordcloud}(
  slise,
  title = NULL,
  labels = c("Low", "High"),
  treshold = 1e-08,
  local = TRUE,
  ...
)
}
\arguments{
\item{slise}{The slise object}

\item{title}{The title of the plot (may include a `\%s`, which will be replaced by the prediction)}

\item{labels}{The class labels (vector with two strings: c(y_low, y_high), default: c("Low", "High"))}

\item{treshold}{Treshold for ignored value (default: 1e-8)}

\item{local}{Only display the words relevant for the explained item (default: TRUE)}

\item{...}{Ignored parameters}
}
\value{
plot
}
\description{
Plot the robust regression or explanation from slise as a wordcloud
}
