% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.slise}
\alias{predict.slise}
\title{Predict with a SLISE}
\usage{
\method{predict}{slise}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{SLISE object}

\item{newdata}{data matrix}
}
\value{
prediction vector
}
\description{
Predict with a SLISE
}
