% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_identity}
\alias{scale_identity}
\title{Create a scale list without any changes}
\usage{
scale_identity(x)
}
\arguments{
\item{x}{the vector to (not) scale}
}
\value{
a list(scaled = x, center = 0, scale = 1)
}
\description{
Create a scale list without any changes
}
