% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_one_range}
\alias{scale_one_range}
\title{Scale a vector to a range of approx one}
\usage{
scale_one_range(x)
}
\arguments{
\item{x}{the vector to normalise}
}
\value{
a list(scaled, center, scale)
}
\description{
Scale a vector to a range of approx one
}
