% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_robust}
\alias{scale_robust}
\title{Robust Scale
A scale that can handle zero variance}
\usage{
scale_robust(x, center = TRUE, scale = TRUE, remove_constant = TRUE)
}
\arguments{
\item{x}{the vector/matrix to normalise}

\item{center}{Should constant columns be centered (TRUE)}

\item{scale}{Should constant columns be scaled (TRUE)}

\item{remove_constant}{Should constant columns be removed (TRUE)}
}
\value{
a list(scaled, center, scale, mask)
}
\description{
Robust Scale
A scale that can handle zero variance
}
