% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{simple_pca}
\alias{simple_pca}
\title{Calculate the PCA rotation matrix
The implementation is based on stats::prcomp.
Assumes the data has already been centered and scaled (if that is desired).}
\usage{
simple_pca(X, dimensions, tolerance = 1e-10)
}
\arguments{
\item{X}{the matrix to reduce}

\item{dimensions}{the number of dimensions after PCA}

\item{tolerance}{remove components with variance less than the tolerance}
}
\value{
pca rotation matrix
}
\description{
Calculate the PCA rotation matrix
The implementation is based on stats::prcomp.
Assumes the data has already been centered and scaled (if that is desired).
}
