% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\docType{package}
\name{slise-package}
\alias{slise}
\alias{slise-package}
\title{slise: Sparse Linear Subset Explanations}
\description{
An implementation of the SLISE
    algorithm for robust regression and
    explaining outcomes from black box models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edahelsinki/slise}
}

}
\author{
\strong{Maintainer}: Anton Björklund \email{anton.bjorklund@helsinki.fi}

}
