% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{slise.explain_comb}
\alias{slise.explain_comb}
\title{SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with sparsity from a combinatorial search rather than Lasso!}
\usage{
slise.explain_comb(X, Y, x, ..., variables = 4)
}
\arguments{
\item{X}{matrix of independent variables}

\item{Y}{vector of the dependent variable}

\item{x}{the sample to be explained}

\item{...}{other parameters to slise.explain}

\item{variables}{the number of non-zero coefficients}
}
\value{
SLISE object
}
\description{
SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with sparsity from a combinatorial search rather than Lasso!
}
