% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.formula}
\alias{slise.formula}
\title{SLISE for robust regression (using a formula).}
\usage{
slise.formula(
  formula,
  data,
  epsilon,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  normalise = FALSE,
  initialisation = slise_initialisation_candidates,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data for the formula}

\item{epsilon}{Error tolerance}

\item{lambda1}{L1 regularisation coefficient (default: 0)}

\item{lambda2}{L2 regularisation coefficient (default: 0)}

\item{weight}{Optional weight vector (default: NULL)}

\item{normalise}{Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)}

\item{initialisation}{Function that gives the initial alpha and beta, or a list containing the initial alpha and beta (default: slise_initialisation_candidates)}

\item{...}{
  Arguments passed on to \code{\link[=graduated_optimisation]{graduated_optimisation}}, \code{\link[=slise_initialisation_candidates]{slise_initialisation_candidates}}
  \describe{
    \item{\code{beta_max}}{Stopping sigmoid steepness (default: 20 / epsilon^2)}
    \item{\code{max_approx}}{Approximation ratio when selecting the next beta (default: 1.15)}
    \item{\code{max_iterations}}{Maximum number of OWL-QN iterations (default: 300)}
    \item{\code{debug}}{Should debug statement be printed each iteration (default: FALSE)}
    \item{\code{num_init}}{the number of initial subsets to generate (default: 500)}
    \item{\code{beta_max_init}}{the maximum sigmoid steepness in the initialisation}
    \item{\code{pca_treshold}}{the maximum number of columns without using PCA (default: 10)}
  }}
}
\value{
slise.object
}
\description{
It is highly recommended that you normalise the data,
either before using SLISE or by setting normalise = TRUE.
}
\examples{
data <- data.frame(y = rnorm(8), a = rnorm(8), b = rnorm(8))
model <- slise.formula(y ~ a * b + abs(a), data, 0.1, normalise = TRUE)
}
