% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.object}
\alias{slise.object}
\title{Create a result object for SLISE that is similar to other regression method results}
\usage{
slise.object(
  alpha,
  X,
  Y,
  epsilon,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  intercept = FALSE,
  logit = FALSE,
  x = NULL,
  y = NULL,
  ...
)
}
\arguments{
\item{alpha}{linear model}

\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{lambda1}{L1 regularisation coefficient (default: 0)}

\item{lambda2}{L2 regularisation coefficient (default: 0)}

\item{weight}{weight vector (default: NULL)}

\item{intercept}{does the model have an intercept (default: FALSE)}

\item{logit}{has the target been logit-transformed (default: FALSE)}

\item{x}{explained item x (default: NULL)}

\item{y}{explained item y (default: NULL)}

\item{...}{other variables to add to the SLISE object}
}
\value{
list(coefficients=unscale(alpha), X, Y, scaled=data, lambda1, lambda2, alpha, subset=[r_i<epsilon], value=loss, epsilon, loss, weight, ...)
}
\description{
Create a result object for SLISE that is similar to other regression method results
}
