% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.object_unnormalise}
\alias{slise.object_unnormalise}
\title{Turn a `slise.object` result based on normalised data to a `slise.object` result with unnormalised data.
The normalised results are retained, but with a 'normalised_' prefix.}
\usage{
slise.object_unnormalise(object, X, Y, x = NULL, y = NULL)
}
\arguments{
\item{object}{output from `slise.object`}

\item{X}{unnormalised data matrix}

\item{Y}{unnormalised response vector}

\item{x}{explained item x (default: NULL)}

\item{y}{explained item y (default: NULL)}
}
\value{
object with SLISE results
}
\description{
Turn a `slise.object` result based on normalised data to a `slise.object` result with unnormalised data.
The normalised results are retained, but with a 'normalised_' prefix.
}
