% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.preprocess}
\alias{slise.preprocess}
\title{Preprocess the data as necessary before running SLISE}
\usage{
slise.preprocess(
  X,
  Y,
  epsilon,
  x = NULL,
  y = NULL,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  intercept = FALSE,
  normalise = FALSE,
  logit = FALSE
)
}
\arguments{
\item{X}{Matrix of independent variables}

\item{Y}{Vector of the target variable}

\item{epsilon}{Error tolerance}

\item{x}{The sample to be explained (or index if y is null)}

\item{y}{The prediction to be explained (default: NULL)}

\item{lambda1}{L1 regularisation coefficient (default: 0)}

\item{lambda2}{L2 regularisation coefficient (default: 0)}

\item{weight}{Optional weight vector (default: NULL)}

\item{intercept}{Should an intercept be added (default: TRUE)}

\item{normalise}{Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)}

\item{logit}{Logit transform Y from probabilities to real values (default: FALSE)}
}
\value{
list(X_orig, Y_orig, X, Y, x_orig, y_orig, x, y)
}
\description{
Preprocess the data as necessary before running SLISE
}
