% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{slise.raw}
\alias{slise.raw}
\title{SLISE Regression
The raw interface for SLISE, initialisation and scaling
  (including locality) has to be done in advance.
This function essentially wraps graduated_optimisation
  and create_slise.}
\usage{
slise.raw(X, Y, alpha = rep(0, ncol(X)), epsilon = 0.1, lambda = 0,
  beta = 0, ...)
}
\arguments{
\item{X}{matrix of independent variables}

\item{Y}{vector of the response variable}

\item{alpha}{Starting alpha}

\item{epsilon}{error tolerance (will be scaled to represent a percentage, e.g. 0.1 == 10\%)}

\item{lambda}{sparsity reguraliser}

\item{beta}{Starting sigmoid steepness}

\item{...}{other parameters to the optimiser}

\item{max_approx}{Target approximation ratio for selecting graduated optimisation step size (1.2)}

\item{beta_max}{Stopping sigmoid steepness (25)}

\item{max_iterations}{Maximum number of OWL-QN steps per graduated optimisation step (250)}
}
\value{
slise object (coefficients, subset, value, X, Y, lambda, epsilon, scaled, alpha)
}
\description{
SLISE Regression
The raw interface for SLISE, initialisation and scaling
  (including locality) has to be done in advance.
This function essentially wraps graduated_optimisation
  and create_slise.
}
