% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{slise_initialisation_lasso}
\alias{slise_initialisation_lasso}
\title{Initialise the graduated optimisation with a LASSO solution}
\usage{
slise_initialisation_lasso(
  X,
  Y,
  weight = NULL,
  lambda1 = 0,
  max_iterations = 300,
  ...
)
}
\arguments{
\item{X}{data matrix}

\item{Y}{response vector}

\item{weight}{weight vector (default: NULL)}

\item{lambda1}{L1 regularisation (default: 0)}

\item{max_iterations}{if ncol(X) is huge, then ols is replaced with optimisation (default:300)}

\item{...}{unused parameters}
}
\value{
list(alpha, beta)
}
\description{
Initialise the graduated optimisation with a LASSO solution
}
