% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparsity}
\alias{sparsity}
\title{Sparsity
Count the non-zero coefficients}
\usage{
sparsity(x, treshold = 0)
}
\arguments{
\item{x}{vector}

\item{treshold}{treshold for approximately zero (0)}
}
\value{
number of non-zero values
}
\description{
Sparsity
Count the non-zero coefficients
}
