% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{unscale}
\alias{unscale}
\title{Unscale a scaled matrix / vector}
\usage{
unscale(x, scaled = NULL)
}
\arguments{
\item{x}{the matrix / vector to unscale}

\item{scaled}{optional object with "scaled:..." attributes}
}
\value{
x unscaled
}
\description{
Unscale a scaled matrix / vector
}
