% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_min_n}
\alias{which_min_n}
\title{Which min n
Get the indecies of the n smallest values using partial sort}
\usage{
which_min_n(x, n)
}
\arguments{
\item{x}{vector}

\item{n}{the number of indices}
}
\value{
vector of indecies
}
\description{
Which min n
Get the indecies of the n smallest values using partial sort
}
