% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{graduated_optimisation}
\alias{graduated_optimisation}
\title{Graduated Optimisation to solve the SLISE problem}
\usage{
graduated_optimisation(alpha, X, Y, epsilon = 0.1, lambda = 0,
  beta = 0, beta_max = 25, max_approx = 1.2, max_iterations = 100,
  debug = FALSE, ..., beta_min_increase = beta_max * 5e-04,
  beta_start_max = NULL)
}
\arguments{
\item{alpha}{initial linear model (if NULL then OLS)}

\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{lambda}{L1 coefficient (0)}

\item{beta}{starting sigmoid steepness (0 => convex problem)}

\item{beta_max}{stopping sigmoid steepness (25)}

\item{max_approx}{approximation ratio when selecting the next beta (1.2)}

\item{max_iterations}{maximum number of OWL-QN iterations (100)}

\item{debug}{should debug statement be printed each iteration (FALSE)}

\item{...}{Additional parameters to OWL-QN}

\item{beta_min_increase}{the minimum increase of beta each iteration (beta_max * 0.0005)}

\item{beta_start_max}{Ignored}
}
\value{
lbfgs object with beta (max) and the number of iteration steps
}
\description{
Graduated Optimisation to solve the SLISE problem
}
