% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{graduated_optimisation}
\alias{graduated_optimisation}
\title{Graduated Optimisation to solve the SLISE problem}
\usage{
graduated_optimisation(
  alpha,
  X,
  Y,
  epsilon,
  beta = 0,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  beta_max = 20/epsilon^2,
  max_approx = 1.15,
  max_iterations = 300,
  beta_min_increase = beta_max * 5e-04,
  debug = FALSE,
  ...
)
}
\arguments{
\item{alpha}{Initial linear model (if NULL then OLS)}

\item{X}{Data matrix}

\item{Y}{Response vector}

\item{epsilon}{Error tolerance}

\item{beta}{Starting sigmoid steepness (default: 0 == convex problem)}

\item{lambda1}{L1 coefficient (default: 0)}

\item{lambda2}{L1 coefficient (default: 0)}

\item{weight}{Weight vector (default: NULL == no weights)}

\item{beta_max}{Stopping sigmoid steepness (default: 20 / epsilon^2)}

\item{max_approx}{Approximation ratio when selecting the next beta (default: 1.15)}

\item{max_iterations}{Maximum number of OWL-QN iterations (default: 300)}

\item{beta_min_increase}{Minimum amount to increase beta (default: beta_max * 0.0005)}

\item{debug}{Should debug statement be printed each iteration (default: FALSE)}

\item{...}{Additional parameters to OWL-QN}
}
\value{
lbfgs object with beta (max) and the number of iteration steps
}
\description{
Graduated Optimisation to solve the SLISE problem
}
