% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_sharp}
\alias{loss_sharp}
\title{Sharp Loss Function
Exact loss function without gradients}
\usage{
loss_sharp(alpha, X, Y, epsilon, lambda1 = 0, lambda2 = 0, weight = NULL)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{X}{The data matrix}

\item{Y}{The response vector}

\item{epsilon}{The acceptable error}

\item{lambda1}{The L1 regulariser (default: 0)}

\item{lambda2}{The L2 regulariser (default: 0)}

\item{weight}{weight vector (default: NULL)}
}
\value{
The loss value
}
\description{
Sharp Loss Function
Exact loss function without gradients
}
