% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{loss_smooth}
\alias{loss_smooth}
\title{Smooth Loss
A loss function for when you want gradients}
\usage{
loss_smooth(
  alpha,
  X,
  Y,
  epsilon,
  beta,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL
)
}
\arguments{
\item{alpha}{The vector to calculate loss for}

\item{X}{The data matrix}

\item{Y}{The response vector}

\item{epsilon}{The acceptable error}

\item{beta}{The steepness of the sigmoid}

\item{lambda1}{The L1 regulariser (default: 0)}

\item{lambda2}{The L2 regulariser (default: 0)}

\item{weight}{weight vector (default: NULL)}
}
\value{
The loss value
}
\description{
Smooth Loss
A loss function for when you want gradients
}
