% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{matching_epsilon}
\alias{matching_epsilon}
\title{Find the matching *epsilon}
\usage{
matching_epsilon(residuals2, epsilon2, beta, weight = NULL)
}
\arguments{
\item{residuals2}{squared residuals}

\item{epsilon2}{squared error tolerance}

\item{beta}{sigmoid steepness}

\item{weight}{weight vector (default: NULL)}
}
\value{
*epsilon
}
\description{
Find the matching *epsilon
}
