% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{owlqn_r}
\alias{owlqn_r}
\title{OWL-QN for optimising loss_smooth
R implementation}
\usage{
owlqn_r(
  alpha,
  X,
  Y,
  epsilon,
  beta,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  max_iterations = 300,
  ...,
  invisible = TRUE
)
}
\arguments{
\item{alpha}{linear model to optimise}

\item{X}{data matrix}

\item{Y}{response vector}

\item{epsilon}{error tolerance}

\item{beta}{sigmoid steepness}

\item{lambda1}{L1 coefficient (default: 0)}

\item{lambda2}{L1 coefficient(default: 0)}

\item{weight}{weight vector (default: NULL)}

\item{max_iterations}{number of OWL-QN iterations (default: 300)}

\item{...}{other parameters to OWL-QN}

\item{invisible}{no terminal output (default: TRUE)}
}
\value{
lbfgs object
}
\description{
OWL-QN for optimising loss_smooth
R implementation
}
