% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slise}
\alias{plot.slise}
\title{Plot the robust regression or explanation from slise}
\usage{
\method{plot}{slise}(slise, cols = 1, title = "SLISE", labels = NULL,
  other = NULL, threed = FALSE, ...)
}
\arguments{
\item{slise}{The slise object}

\item{cols}{The columns in the data to plot}

\item{title}{(Optional) The title of the plot (and result names when using other)}

\item{labels}{(Optional) The labels for the x, y, and legend (in that order, can be partial)}

\item{other}{list of other slise objects to include in the plot}

\item{threed}{plot in 3D with two columns}

\item{...}{not used}
}
\description{
Plot the robust regression or explanation from slise
}
\examples{
data <- matrix(rnorm(200), 100, 2)
response <- rnorm(100)
slise <- slise.fit(data, response, epsilon=0.1)
plot(slise, 1:2, threed = TRUE)
}
