% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.slise}
\alias{predict.slise}
\title{Predict with a SLISE}
\usage{
\method{predict}{slise}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{SLISE object}

\item{newdata}{data matrix}

\item{...}{not used}
}
\value{
prediction vector
}
\description{
Predict with a SLISE
}
\examples{
X <- matrix(rnorm(200), 100, 2)
Y <- rnorm(100)
index <- 10
model <- slise.explain(X, Y, index)
prediction <- predict(model, X)
}
