% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{predict.slise}
\alias{predict.slise}
\title{Predict with a SLISE}
\usage{
\method{predict}{slise}(object, newdata = NULL, ..., logit = FALSE)
}
\arguments{
\item{object}{SLISE object}

\item{newdata}{data matrix}

\item{...}{ignored additional parameters}

\item{logit}{return the result in logit space if a logit has been applied to Y (default: FALSE)}
}
\value{
prediction vector
}
\description{
Predict with a SLISE
}
