% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanation.R
\name{print.slise}
\alias{print.slise}
\title{Print the robust regression or explanation from slise}
\usage{
\method{print}{slise}(slise, title = "SLISE", ...)
}
\arguments{
\item{slise}{The slise object}

\item{title}{(Optional) The title of the result}
}
\description{
Print the robust regression or explanation from slise
}
\examples{
print(slise.fit(data, response, epsilon=0.1, variables=5))
}
