% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_identity}
\alias{scale_identity}
\title{A variant of `scale` that only adds the attributes}
\usage{
scale_identity(x)
}
\arguments{
\item{x}{the vector to (not) scale}
}
\value{
x (with attributes "scaled:center" and "scaled:scale")
}
\description{
A variant of `scale` that only adds the attributes
}
