% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scale_robust}
\alias{scale_robust}
\title{Robust Scale
A variant of 'scale' that is based on median and mad (instead of mean and sd).
It can handle zero variance without producing nan:s.}
\usage{
scale_robust(x, th = .Machine$double.eps)
}
\arguments{
\item{x}{the vector/matrix to normalise}

\item{th}{threshold for the scale being zero}
}
\value{
scaled_x (with attributes "scaled:center" and "scaled:scale")
}
\description{
Robust Scale
A variant of 'scale' that is based on median and mad (instead of mean and sd).
It can handle zero variance without producing nan:s.
}
