% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp.R
\docType{package}
\name{slise-package}
\alias{slise}
\alias{slise-package}
\title{slise: Sparse Linear Subset Explanations}
\description{
An implementation of the SLISE algorithm (for robust regression and explaining outcomes from black box models).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edahelsinki/slise}
}

}
\author{
\strong{Maintainer}: Anton Björklund \email{anton.bjorklund@helsinki.fi} (\href{https://orcid.org/0000-0002-7749-2918}{ORCID})

Authors:
\itemize{
  \item Andreas Henelius (\href{https://orcid.org/0000-0002-4040-6967}{ORCID})
  \item Kai Puolamäki (\href{https://orcid.org/0000-0003-1819-1047}{ORCID})
}

}
