% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.explain_comb}
\alias{slise.explain_comb}
\title{SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with sparsity from a combinatorial search rather than Lasso!}
\usage{
slise.explain_comb(X, Y, epsilon, x, y = NULL, ..., variables = 4)
}
\arguments{
\item{X}{matrix of independent variables}

\item{Y}{vector of the dependent variable}

\item{epsilon}{error tolerance}

\item{x}{the sample to be explained (or index if y is null)}

\item{y}{the prediction to be explained}

\item{...}{
  Arguments passed on to \code{\link[=slise.explain]{slise.explain}}
  \describe{
    \item{\code{lambda1}}{L1 regularisation coefficient (default: 0)}
    \item{\code{lambda2}}{L2 regularisation coefficient (default: 0)}
    \item{\code{weight}}{Optional weight vector (default: NULL)}
    \item{\code{normalise}}{Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)}
    \item{\code{logit}}{Logit transform Y from probabilities to real values (default: FALSE)}
    \item{\code{initialisation}}{function that gives the initial alpha and beta, or a list containing the initial alpha and beta (default: slise_initialisation_candidates)}
  }}

\item{variables}{the number of non-zero coefficients}
}
\value{
SLISE object
}
\description{
SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with sparsity from a combinatorial search rather than Lasso!
}
