% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{slise.explain_find}
\alias{slise.explain_find}
\title{SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with a binary search for sparsity!}
\usage{
slise.explain_find(..., lambda = 5, variables = 4, iters = 10,
  treshold = 1e-04)
}
\arguments{
\item{...}{parameters to slise.explain}

\item{lambda}{the starting value of the search}

\item{variables}{number of non-zero coefficients}

\item{iters}{number of search iterations}

\item{treshold}{treshold for zero coefficient}
}
\value{
SLISE object
}
\description{
SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with a binary search for sparsity!
}
\examples{
X <- matrix(rnorm(800), 100, 8)
Y <- rnorm(100)
index <- 10
model <- slise.explain_find(X, Y, index, variables = 4)
}
