% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.explain_find}
\alias{slise.explain_find}
\title{SLISE Black Box Explainer
Use SLISE for explaining predictions made by a black box.
BUT with a binary search for sparsity!}
\usage{
slise.explain_find(
  ...,
  lambda1 = 5,
  variables = 4,
  iters = 10,
  treshold = 1e-04
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=slise.explain]{slise.explain}}
  \describe{
    \item{\code{X}}{Matrix of independent variables}
    \item{\code{Y}}{Vector of the dependent variable}
    \item{\code{epsilon}}{Error tolerance}
    \item{\code{x}}{The sample to be explained (or index if y is null)}
    \item{\code{y}}{The prediction to be explained (default: NULL)}
    \item{\code{lambda2}}{L2 regularisation coefficient (default: 0)}
    \item{\code{weight}}{Optional weight vector (default: NULL)}
    \item{\code{normalise}}{Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)}
    \item{\code{logit}}{Logit transform Y from probabilities to real values (default: FALSE)}
    \item{\code{initialisation}}{function that gives the initial alpha and beta, or a list containing the initial alpha and beta (default: slise_initialisation_candidates)}
  }}

\item{lambda1}{the starting value of the search}

\item{variables}{number of non-zero coefficients}

\item{iters}{number of search iterations}

\item{treshold}{treshold for zero coefficient}
}
\value{
SLISE object
}
\description{
DEPRECATED: This is a simple binary search, no need for a separate function
}
