% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slise.R
\name{slise.fit}
\alias{slise.fit}
\title{SLISE for robust regression.}
\usage{
slise.fit(
  X,
  Y,
  epsilon,
  lambda1 = 0,
  lambda2 = 0,
  weight = NULL,
  intercept = TRUE,
  normalise = FALSE,
  initialisation = slise_initialisation_candidates,
  ...
)
}
\arguments{
\item{X}{Matrix of independent variables}

\item{Y}{Vector of the response variable}

\item{epsilon}{Error tolerance}

\item{lambda1}{L1 regularisation coefficient (default: 0)}

\item{lambda2}{L2 regularisation coefficient (default: 0)}

\item{weight}{Optional weight vector (default: NULL)}

\item{intercept}{Should an intercept be added (default: TRUE)}

\item{normalise}{Preprocess X and Y by scaling, note that epsilon is not scaled (default: FALSE)}

\item{initialisation}{Function that gives the initial alpha and beta, or a list containing the initial alpha and beta (default: slise_initialisation_candidates)}

\item{...}{
  Arguments passed on to \code{\link[=graduated_optimisation]{graduated_optimisation}}, \code{\link[=slise_initialisation_candidates]{slise_initialisation_candidates}}
  \describe{
    \item{\code{beta_max}}{Stopping sigmoid steepness (default: 20 / epsilon^2)}
    \item{\code{max_approx}}{Approximation ratio when selecting the next beta (default: 1.15)}
    \item{\code{max_iterations}}{Maximum number of OWL-QN iterations (default: 300)}
    \item{\code{debug}}{Should debug statement be printed each iteration (default: FALSE)}
    \item{\code{num_init}}{the number of initial subsets to generate (default: 500)}
    \item{\code{beta_max_init}}{the maximum sigmoid steepness in the initialisation}
    \item{\code{pca_treshold}}{the maximum number of columns without using PCA (default: 10)}
  }}
}
\value{
slise.object
}
\description{
It is highly recommended that you normalise the data,
either before using SLISE or by setting normalise = TRUE.
}
\examples{
X <- matrix(rnorm(32), 8, 4)
Y <- rnorm(8)
model <- slise.fit(X, Y, (max(Y) - min(Y)) * 0.1)
predicted <- predict(model, X)
}
